Code/requirements.txt
Beaker==1.5.4
Brlapi==0.6.3
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
Pillow==2.6.2
PyYAML==3.11
Pygments==1.6
backports-abc==0.4
bottle==0.12.6
chardet==2.2.1
colorlover==0.2.1
cufflinks==0.7.2
cupshelpers==1.0
decorator==3.4.0
docutils==0.12
graphviz==0.4.10
ipdb==0.8
ipykernel==4.3.1
ipyparallel==5.0.1
ipython==4.1.2
ipython-genutils==0.1.0
ipywidgets==4.1.1
jsonpointer==1.0
jsonschema==2.3.0
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.1.1
jupyter-core==4.0.6
keyring==5.0
matplotlib==1.3.1
mistune==0.7.2
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.11
nltk==3.0.0
nose==1.3.7
notebook==4.1.0
numpy==1.8.2
pandas==0.18.0
pandocfilters==1.2.4
path.py==8.1.2
pexpect==3.1
pickleshare==0.6
plotly==1.9.7
ptyprocess==0.5.1
pycups==1.9.70
pycurl==7.19.3.1
pygobject==3.14.0
pygraphviz==1.3.1
pyparsing==2.0.3
python-dateutil==2.2
pytz==2012d
pyxdg==0.25
pyzmq==14.3.1
qtconsole==4.2.0
requests==2.7.0
scipy==0.14.1
simplegeneric==0.8.1
six==1.9.0
terminado==0.6
tornado==4.3
traitlets==4.1.0
urllib3==1.10.4
virtualenv==12.0.7
wordcloud==1.2.1
